+++ lib/e_menu/src/apavedialog.tcl
@@ -431,8 +431,7 @@ oo::class create ::apave::APaveDialog {
       incr lto
       lassign [my GetLinePosition $txt $lfrom.0] linestart lineend
       set duptext [$txt get $linestart $lineend]
-      $txt configure -autoseparators no
-      $txt edit separator
+      ::apave::undoIn $txt
       $txt delete $linestart $lineend
       $txt insert $lto.0 $duptext
       ::tk::TextSetCursor $txt [NewRow $pos $to]
@@ -442,8 +441,7 @@ oo::class create ::apave::APaveDialog {
       if {[lsearch -glob [$txt tag names] tagCOM*]>-1} {
         catch {::hl_tcl::my::Modified $txt insert $lto.0 $lto.end}
       }
-      $txt edit separator
-      $txt configure -autoseparators yes
+      ::apave::undoOut $txt
       if {$dobreak} {return -code break}
     }
     return