+++ lib/e_menu/src/baltip/pkgIndex.tcl
@@ -1,4 +1,4 @@
-package ifneeded baltip 1.4.0 [list source [file join $dir baltip.tcl]]
+package ifneeded baltip 1.4.1 [list source [file join $dir baltip.tcl]]
 
 namespace eval ::baltip {
   variable _ruff_preamble {
@@ -110,11 +110,23 @@ To get all or specific settings of *baltip*:
       # or this way:
       ::baltip cget ?-option?
 
+To get a specific widget's tip option:
+
+      ::baltip::cget widgetpath -option
+      # or this way:
+      ::baltip cget widgetpath -option
+
 To set some options:
 
       ::baltip::configure -option value ?-option value?
       # or this way:
       ::baltip config -option value ?-option value?
+
+To set a specific widget's tip option:
+
+      ::baltip::configure widgetpath -option value ?-option value?
+      # or this way:
+      ::baltip config widgetpath -option value ?-option value?
 
 **Note**: the options set with `configure` command are *global*, i.e. active for all tips.
 The options set with `tip` command are *local*, i.e. active for the specific tip.